Conflicts:
	xcloud-iam-common/src/main/java/com/wl4g/iam/common/config/AbstractIamConfiguration.java

+++ xcloud-iam-core/src/main/java/com/wl4g/iam/service/GroupService.java
@@ -31,12 +31,12 @@ public interface GroupService {
 
 	void save(Group group);
 
-	void del(Integer id);
+	void del(Long id);
 
-	Group detail(Integer id);
+	Group detail(Long id);
 
 	Set<Group> getGroupsSet(User user);
 
-	Group getParent(List<Group> groups, Integer parentId);
+	Group getParent(List<Group> groups, Long parentId);
 
 }
